solucionado tema login, se crean mal los dates

+++ service-dateEntity/src/main/java/com/example/service_dateEntity/model/dtos/RequestCreateAvailabilityDto.java
@@ -8,7 +8,7 @@ public record RequestCreateAvailabilityDto(
         String employeeId,
         DaysOfTheWeek dayOfTheWeek,
         Integer duration,
-        List<Integer> range,
+        List<Double> range,
         Integer rest
 ) {
 }
